my_test_framework/requirements.txt
selenium==4.25.0
pytest==8.3.3
pytest-html==4.1.1
PyYAML==6.0.2